Further conversion of int(32) -> size_t for size-related comm arguments

+++ runtime/include/chpl-comm.h
@@ -81,16 +81,16 @@ extern const int chpl_heterogeneous;
 // wait for the GET to complete. The destination buffer must not be modified
 // before the request completes (after waiting on the returned handle)
 chpl_comm_nb_handle_t chpl_comm_get_nb(void* addr, c_nodeid_t node, void* raddr,
-                                     int32_t elemSize, int32_t typeIndex,
-                                     int32_t len,
+                                     size_t elemSize, int32_t typeIndex,
+                                     size_t len,
                                      int ln, c_string fn);
 
 // Do a PUT in a nonblocking fashion, returning a handle which can be used to
 // wait for the PUT to complete. The source buffer must not be modified before
 // the request completes (after waiting on the returned handle)
 chpl_comm_nb_handle_t chpl_comm_put_nb(void *addr, c_nodeid_t node, void* raddr,
-                                    int32_t elemSize, int32_t typeIndex,
-                                    int32_t len,
+                                    size_t elemSize, int32_t typeIndex,
+                                    size_t len,
                                     int ln, c_string fn);
 
 // Returns nonzero iff the handle has already been waited for and has
@@ -215,7 +215,7 @@ void chpl_comm_broadcast_global_vars(int numGlobals);
 // currently in use on any platforms, but is being retained in the
 // event that we wish to re-enable this capability in the future.
 // 
-void chpl_comm_broadcast_private(int id, int32_t size, int32_t tid);
+void chpl_comm_broadcast_private(int id, size_t size, int32_t tid);
 
 //
 // Barrier for synchronization between all top-level locales; currently
@@ -261,7 +261,7 @@ void chpl_comm_exit(int all, int status);
 //   size and locale are part of p
 //
 void  chpl_comm_put(void* addr, c_nodeid_t node, void* raddr,
-                    int32_t elemSize, int32_t typeIndex, int32_t len,
+                    size_t elemSize, int32_t typeIndex, size_t len,
                     int ln, c_string fn);
 
 //
@@ -272,7 +272,7 @@ void  chpl_comm_put(void* addr, c_nodeid_t node, void* raddr,
 //   size and locale are part of p
 //
 void  chpl_comm_get(void *addr, c_nodeid_t node, void* raddr,
-                    int32_t elemSize, int32_t typeIndex, int32_t len,
+                    size_t elemSize, int32_t typeIndex, size_t len,
                     int ln, c_string fn);
 
 //
@@ -289,7 +289,7 @@ void  chpl_comm_get(void *addr, c_nodeid_t node, void* raddr,
 //
 void  chpl_comm_put_strd(void* dstaddr, void* dststrides, int32_t dstlocale, 
                      void* srcaddr, void* srcstrides, void* count,
-                     int32_t stridelevels, int32_t elemSize, int32_t typeIndex, 
+                     size_t stridelevels, size_t elemSize, int32_t typeIndex, 
                      int ln, c_string fn);
 
 //
@@ -297,7 +297,7 @@ void  chpl_comm_put_strd(void* dstaddr, void* dststrides, int32_t dstlocale,
 //
 void  chpl_comm_get_strd(void* dstaddr, void* dststrides, int32_t srclocale, 
                      void* srcaddr, void* srcstrides, void* count,
-                     int32_t stridelevels, int32_t elemSize, int32_t typeIndex, 
+                     size_t stridelevels, size_t elemSize, int32_t typeIndex, 
                      int ln, c_string fn);
 
 //
@@ -308,7 +308,7 @@ void  chpl_comm_get_strd(void* dstaddr, void* dststrides, int32_t srclocale,
 // The local char[] buffer is leaked. :(
 //
 void chpl_gen_comm_wide_string_get(void* addr,
-  c_nodeid_t node, void* raddr, int32_t elemSize, int32_t typeIndex, int32_t len,
+  c_nodeid_t node, void* raddr, size_t elemSize, int32_t typeIndex, size_t len,
                                    int ln, c_string fn);
 
 //
@@ -318,19 +318,19 @@ void chpl_gen_comm_wide_string_get(void* addr,
 //   multiple forks to the same locale should be handled concurrently
 //
 void chpl_comm_fork(c_nodeid_t node, c_sublocid_t subloc,
-                    chpl_fn_int_t fid, void *arg, int32_t arg_size);
+                    chpl_fn_int_t fid, void *arg, size_t arg_size);
 
 //
 // non-blocking fork
 //
 void chpl_comm_fork_nb(c_nodeid_t node, c_sublocid_t subloc,
-                       chpl_fn_int_t fid, void *arg, int32_t arg_size);
+                       chpl_fn_int_t fid, void *arg, size_t arg_size);
 
 //
 // fast (non-forking) fork (i.e., run in handler)
 //
 void chpl_comm_fork_fast(c_nodeid_t node, c_sublocid_t subloc,
-                         chpl_fn_int_t fid, void *arg, int32_t arg_size);
+                         chpl_fn_int_t fid, void *arg, size_t arg_size);
 
 
 //
